refactor(mobile): split calendar store
- extract session and state modules
- add store source-contract test
